inference_service/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
mlflow
Pillow
numpy
anthropic
